+++ ChangeLog.md
@@ -4,6 +4,8 @@ Follow the principles in [keepachangelog.com](https://keepachangelog.com)!
 
 # v.Next (Current)
 
+# [v0.2.0-beta.62](https://github.com/moxb/moxb/releases/tag/v0.2.0-beta.62) (2021-02-02)
+
 - OneOfAnt: add support for vertical layout of choices
 - OneOfAnt: add support for displaying explanation besides the choices